requirements.txt
django
python-dotenv
django-crispy-forms
djangorestframework
django-cors-headers

